New API: Tidy up and address review actions
Consolidated LSPExecutor and LanguageServers class into a single class
Removed accidental changes whereby some references to LanguageServerWrapper had been replaced by ILSWrapper by the automated refactoring
Changed ILSWrapper to ILanguageServerWrapper to be more style conformant
Added missing copyright/license headers
